Patch
Добавлено:
- "Бустеры" для оружия
- "Мультишот" - дает дополнительно по 1 снаряду от основного направления выстрела
- "Слэш-урон" - снаряды получают дополнительно зону поражения
- "Буст скорострельности" - увеличивает скорострельность
- "Буст урона" - увеличивает урон
- Боковая часть стены за уровнем теперь подвергается туману

Изменено:
- Теперь с минибоссов выпадают только "Бустеры"
- Правила заполнения дропа с мобов
- В ячейки дропа можно добавлять список предметов
- В режиме "OneItem" выпадет только 1 предмет из списка

Исправлено:
- Текст "начисления очков" теперь имеет шрифт варианта перевода

+++ RavenKnight/Assets/Project/Scripts/Scene/GameScene/UI/UI_Score.cs
@@ -45,8 +45,8 @@ public class UI_Score : MonoBehaviour
     {
         if (item.TypePickupItem != pickupItem) return;
 
-        score += item.Value;
-        OnCreateScoreText(item.transform.position, item.Value);
+        score += (int)item.Value;
+        OnCreateScoreText(item.transform.position, (int)item.Value);
 
         string scoreString = "";
         for (int i = score.ToString().Length; i < 7; i++)
